vector_retrieval/requirements.txt
groq
sentence-transformers
faiss-cpu
numpy
matplotlib
huggingface_hub
python-dotenv